requirements.txt
aniso8601==8.1.0
attrs==20.3.0
bson==0.5.10
click==7.1.2
Flask==1.1.2
flask-restplus==0.13.0
importlib-metadata==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
psycopg2-binary
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
Werkzeug==0.16.1
zipp==3.4.0
